Correção de detalhes no autoupdate, e erro no arquivo Update.exe

+++ Update/Update/Program.cs
@@ -4,6 +4,7 @@ using System.Diagnostics;
 using System.IO;
 using System.Linq;
 using System.Text;
+using System.Threading;
 
 namespace Update
 {
@@ -12,14 +13,40 @@ namespace Update
     {
         static void Main(string[] args)
         {
-            var diretorio = args[0].Split('&');
+            string parametros = string.Empty;
+            string[] diretorio = new string[2];
 
-            while (!File.Exists(diretorio[1] + "\\TimesheetNew.exe")) ;
+            Console.WriteLine("Aguardando construção de binarios...");
+            Thread.Sleep(1000);
+            Console.WriteLine("Carregando dados em lote...");
+            Thread.Sleep(2000);
+            Console.WriteLine("Construção em processo...");
 
-            File.Delete(diretorio[0] + "\\Timesheet.exe");
-            File.Move(diretorio[1] + "\\TimesheetNew.exe", diretorio[0] + "\\Timesheet.exe");
+            for (var i = 0; i < args.Length; i++)
+                parametros += args[i] + " ";
 
-            Process.Start(diretorio[0] + "\\Timesheet.exe");
+            parametros = parametros.Substring(parametros.IndexOf("--pathData:"), parametros.Length - parametros.IndexOf("--pathData:"));
+            diretorio = parametros.Trim().Replace("--pathData:", "").Split('|');
+
+            if (!string.IsNullOrEmpty(diretorio[0]))
+            {
+                
+                Console.WriteLine("Dir0: " + diretorio[0]);
+                Console.WriteLine("Dir1: " + diretorio[1]);
+
+                while (!File.Exists(diretorio[1] + "\\TimesheetNew.exe")) ;
+
+                File.Delete(diretorio[0] + "\\Timesheet.exe");
+                File.Move(diretorio[1] + "\\TimesheetNew.exe", diretorio[0] + "\\Timesheet.exe");
+
+                Console.WriteLine("Instalação completa!");
+
+                Process.Start(diretorio[0] + "\\Timesheet.exe");
+                return;
+            }
+
+            Console.WriteLine("Argumentos não foram passados corretamente, entre em contato para mais informações...");
+            
         }
     }
 }
